This commit changes the 'soltesz.py' module into 'moncommands.py' and 'database.py'
Also, findbad*.py include a timeout that should allow them to exit even if the
'futex' bug gets hung somewhere.

Also, the mailer, rt, and others are updated to use monitorconfig.py as the
source of their username and password information rather than random files
here and there.  This also allows us to keep this information out of svn.