One can now specify "sysctl.x.y" attributes where x.y is a kernel variable.
sm.py searches for these slice attributes and adds them to the rspec.

sliver_vs.py processes these in set_resources(), but only if the new
"setup" argument to set_resources() is set to True.  By default
setup=False and should only be set to true when setting up a vserver,
as done in util-vserver-pl/python/vserver.py:__do_chcontext().  The
reason being is that it needs to write /proc/sys/x/y while within the
setup phase of a vserver context when capability checking is disabled.