Tagging module pyplnet - pyplnet-4.2-3
BUGFIX: the "program" argument needs to go with the the m.output() method.  

The bug is that the m.input() method does not take two args and python
will raise an exception to indicate this.  However, the m.input()
method was wrapped in a

try:
	m.input(a,b)
except:
	pass

which masked this error.  The mainly visible side effect is that the
specific modprobe.conf file is not parsed and so any previously
written aliases and options are lost.